fix(eval): fix style exposure attribution and Decimal type issues
- Add .astype(float) at source in data_service.py for load_market_cap/load_pb
  to prevent decimal.Decimal incompatibility with np.log()
- Rewrite attribution.decompose() to handle partial style factor data
  (pre-align panels, allow ≥2 active factors per date)
- Use neutralized factor for attribution when neutralization is enabled
- Fix per_symbol_summary crash when all_syms is empty
- Improve logging for style factor computation failures

+++ backend/factors/riskmodel/liquidity.py
@@ -29,7 +29,7 @@ class Liquidity20d(BaseFactor):
             return pd.DataFrame()
 
         volume = volume.astype(float).sort_index()
-        mktcap = mktcap.reindex(index=volume.index, columns=volume.columns)
+        mktcap = mktcap.astype(float).reindex(index=volume.index, columns=volume.columns)
         turnover = volume / mktcap.replace(0.0, np.nan)
         result = turnover.rolling(window=20, min_periods=10).mean()
         return result.loc[ctx.start_date:]
